Version 0.1.3

+++ XRTK-Core/Assets/XRTK/Inspectors/Utilities/SymbolicLinks/SymbolicLinker.cs
@@ -24,8 +24,8 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
         /// </summary>
         static SymbolicLinker()
         {
+            RunSync(Application.isBatchMode);
             EditorApplication.projectWindowItemOnGUI += OnProjectWindowItemGui;
-            RunSync();
         }
 
         private const string LINK_ICON_TEXT = "<=link=>";
@@ -46,6 +46,8 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
         internal static string ProjectRoot => GitUtilities.RepositoryRootDir;
 
         private static bool isRunningSync;
+
+        public static bool IsSyncing => isRunningSync || MixedRealityPackageUtilities.IsRunningCheck;
 
         /// <summary>
         /// Debug the symbolic linker utility.
@@ -82,7 +84,7 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
         /// <param name="forceUpdate">Bypass the auto load check and force the packages to be updated, even if they're up to date.</param>
         public static async void RunSync(bool forceUpdate = false)
         {
-            if (isRunningSync)
+            if (isRunningSync || EditorApplication.isPlayingOrWillChangePlaymode)
             {
                 return;
             }
@@ -177,9 +179,8 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
             AssetDatabase.Refresh();
 
             EditorApplication.UnlockReloadAssemblies();
-            isRunningSync = false;
-
             MixedRealityPackageUtilities.CheckPackageManifest();
+            isRunningSync = false;
         }
 
         /// <summary>
@@ -205,6 +206,11 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
             sourceRelativePath = sourceRelativePath.Replace(ProjectRoot, string.Empty);
             targetRelativePath = targetRelativePath.Replace(ProjectRoot, string.Empty);
 
+            if (!CreateSymbolicLink(sourceAbsolutePath, targetAbsolutePath))
+            {
+                return;
+            }
+
             SymbolicLink symbolicLink;
 
             // Check if the symbolic link is already registered.
@@ -227,11 +233,6 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
                 {
                     symbolicLink.IsActive = true;
                 }
-            }
-
-            if (!Directory.Exists(targetAbsolutePath))
-            {
-                CreateSymbolicLink(sourceAbsolutePath, targetAbsolutePath);
             }
         }
 
@@ -308,18 +309,18 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
             }
         }
 
-        private static void CreateSymbolicLink(string sourceAbsolutePath, string targetAbsolutePath)
+        private static bool CreateSymbolicLink(string sourceAbsolutePath, string targetAbsolutePath)
         {
             if (string.IsNullOrEmpty(targetAbsolutePath) || string.IsNullOrEmpty(sourceAbsolutePath))
             {
                 Debug.LogError("Unable to create symbolic link with null or empty args");
-                return;
+                return false;
             }
 
             if (!sourceAbsolutePath.Contains(ProjectRoot))
             {
                 Debug.LogError($"The symbolic link you're importing needs to be in your project's git repository.\n{sourceAbsolutePath}\n{ProjectRoot}");
-                return;
+                return false;
             }
 
             targetAbsolutePath = AddSubfolderPathToTarget(sourceAbsolutePath, targetAbsolutePath);
@@ -357,14 +358,15 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
             }
 
             // --------------------> /C mklink /D "C:\Link To Folder" "C:\Users\Name\Original Folder"
-            if (!new Process().Run($"/C mklink /D \"{targetAbsolutePath}\" \"{sourceAbsolutePath}\"", out string error))
+            if (!new Process().Run($"/C mklink /D \"{targetAbsolutePath}\" \"{sourceAbsolutePath}\"", out var error))
             {
                 Debug.LogError($"{error}");
-                return;
+                return false;
             }
 
             Debug.Log($"Successfully created symbolic link to {sourceAbsolutePath}");
             AssetDatabase.Refresh();
+            return true;
         }
 
         private static bool DeleteSymbolicLink(string path)
@@ -377,12 +379,16 @@ namespace XRTK.Inspectors.Utilities.SymbolicLinks
 
             if (new Process().Run($"/C rmdir /q \"{path}\"", out var error))
             {
-                File.Delete($"{path}.meta");
+                if (File.Exists($"{path}.meta"))
+                {
+                    File.Delete($"{path}.meta");
+                }
+
                 AssetDatabase.Refresh();
                 return true;
             }
 
-            Debug.LogError(error);
+            Debug.LogError($"{error}\n{path}");
             return false;
         }
